Move calculator logic to shared libs and components
