+++ app/guide/shared/Calculator/index.ts
@@ -1,0 +1,1 @@
+export * from "./Calculator";
