+++ lib/constants.ts
@@ -2,3 +2,6 @@ export const ONE_SECOND = 1000;
 export const ONE_MINUTE = 60 * ONE_SECOND;
 export const ONE_HOUR = 60 * ONE_MINUTE;
 export const ONE_DAY = 24 * ONE_HOUR;
+
+export const eTradeGainsLossesUrl =
+  "https://us.etrade.com/etx/sp/stockplan#/myAccount/gainsLosses";
